+++ packages/dive/lib/src/dive_format.dart
@@ -6,6 +6,9 @@ class DiveFormat {
   static final formatterShort = DateFormat('mm:ss');
   static final formatterShortest = DateFormat('s.S');
   static final formatterLog = DateFormat('y/M/d H:mm:ss');
+  static final formatterTime = DateFormat('H:mm:ss a');
+  static final formatterRecordingDate = DateFormat('y-MM-dd');
+  static final formatterRecordingTime = DateFormat('HH-mm-ss');
 
   /// For example, for 42 milliseconds
   ///